backend/requirements.txt
# Core Dependencies
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
sqlalchemy==2.0.44

# Payment Processing
stripe==8.0.0

# Azure & Authentication
pymongo==4.6.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
python-dotenv==1.0.0

# Content Sourcing
feedparser==6.0.11
beautifulsoup4==4.14.2
aiohttp==3.13.2
lxml==6.0.2
playwright==1.55.0

# Social Media Sourcers
praw==7.7.1  # Reddit API
ntscraper==0.4.0  # Twitter/X scraping (no API key)
google-api-python-client==2.108.0  # YouTube Data API
newsapi-python==0.2.7  # NewsAPI for news aggregation

# NLP & Keyword Extraction
scikit-learn>=1.3.0
spacy>=3.7.0
yake>=0.4.8
sentence-transformers>=2.2.0

# Sentiment Analysis
vaderSentiment>=3.3.2

# Additional utilities
numpy>=1.24.0

# Development Tools
black==23.12.1
isort==5.13.2
pre-commit==3.6.0

# Note: After installing, download the spaCy language model:
# python -m spacy download en_core_web_md
